Add rollup bundle build
