+++ package.json
@@ -72,8 +72,10 @@
         "friendly-errors-webpack-plugin": "1.6.x",
         "gulp": "github:gulpjs/gulp#4.0",
         "gulp-autoprefixer": "3.1.x",
+        "gulp-better-rollup": "1.1.x",
         "gulp-clean-css": "2.3.x",
         "gulp-inline-ng2-template": "4.0.x",
+        "gulp-ngc": "0.2.x",
         "gulp-rename": "1.2.x",
         "gulp-sourcemaps": "2.4.x",
         "gulp-sass": "3.1.x",
