+++ tsconfig.dev.json
@@ -2,7 +2,6 @@
 	"extends": "./tsconfig.base.json",
 	"compilerOptions": {
 		"outDir": "./build/tsc/app",
-		"module": "es2015",
 		"baseUrl": "",
 		"types": [
 			"node"
